Fix-Refactor-code(test_books)
