+++ app/__init__.py
@@ -26,6 +26,7 @@ def create_app(config_name):
     app.config['JWT_BLACKLIST_TOKEN_CHECKS'] = ['access']
     jwt = JWTManager(app)
 
+    # Disables FlaskSQLAlchemy's event notifications
     app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
     db.init_app(app)
